feat: added watcher and remove css loader; bumped to v0.7.4

+++ tir/__init__.py
@@ -1,9 +1,7 @@
 import logging
 import os
 import shutil
-import time
-
-import sass
+from pathlib import Path
 
 from tir.config import REQUIRED_PATHS, get_config
 from tir.files import hash_content, rm_dir_files
@@ -11,6 +9,7 @@ from tir.posts import Post
 from tir.templates import TemplateLoader
 from tir.tools import is_init, minify_file
 from tir.utils import mktree, scantree
+from tir.watch import watch_multiple_directories
 
 logger = logging.getLogger(__name__)
 
@@ -26,6 +25,7 @@ class Tir(object):
         self.conf = get_config()
 
         self.theme = self.conf.visuals.theme
+        self.theme_path = os.path.join(os.getcwd(), 'layout', self.theme)
         self.lang = self.conf.lang
         self.build_dir = self.conf.build_dir
         self.working_dir = os.getcwd()
@@ -69,30 +69,35 @@ class Tir(object):
         try:
             if os.path.exists(self.build_dir) and not path:
                 shutil.rmtree(self.build_dir)
-            tpl_visuals_dir = os.path.join(os.getcwd(), 'layout', self.theme)
-            print('Selected template:', tpl_visuals_dir)
-            print('Replace assets files')
-            assets_src_dir = os.path.join(tpl_visuals_dir, 'assets')
+            print(f'Selected template: {self.theme} ({self.theme_path})')
+
             assets_target_dir = os.path.join(
                 os.getcwd(), self.build_dir, 'static')
+            assets_src_dir = os.path.join(self.theme_path, 'assets')
 
             tpl_loader = TemplateLoader(
-                layout_directory=tpl_visuals_dir, config=self.conf)
+                layout_directory=self.theme_path, config=self.conf)
 
             print('Building static files...')
             css_dir = os.path.join(assets_target_dir, 'css')
-            mktree(css_dir)
-            compiled_scss = sass.compile(filename=os.path.join(
-                assets_src_dir, 'scss', 'main.scss'))
-            stylesheet_path = os.path.join(css_dir, 'stylesheet.css')
-            with open(stylesheet_path, 'w+', encoding='utf-8') as f:
-                f.write(compiled_scss)
-            minified_stylesheet_path = minify_file(stylesheet_path)
 
-            shutil.copytree(
-                os.path.join(tpl_visuals_dir, 'assets', 'images'),
-                os.path.join(self.build_dir, 'static', 'images')
-            )
+            mktree(css_dir)
+
+            stylesheet_path = os.path.join(css_dir, 'stylesheet.css')
+            main_css_path = Path(assets_src_dir, 'css', 'main.css')
+            minified_stylesheet_path = None
+
+            if main_css_path.exists():
+                shutil.copy(os.path.join(main_css_path), stylesheet_path)
+
+                minified_stylesheet_path = minify_file(stylesheet_path)
+
+                shutil.copytree(
+                    os.path.join(self.theme_path, 'assets', 'images'),
+                    os.path.join(self.build_dir, 'static', 'images'), dirs_exist_ok=True
+                )
+            else:
+                print(f'CSS file {main_css_path} does not exist')
 
             elements = []
 
@@ -112,13 +117,28 @@ class Tir(object):
             raise e
 
     def watch(self):
-        self.oc = hash_content(self.posts_dir)
-        print('Now watching for changes...'.format(self.posts_dir))
-        while 1:
-            time.sleep(2)
-            cc = hash_content(self.posts_dir)
-            if self.oc == cc:
-                pass
-            else:
-                self.oc = cc
-                self.build()
+        directories_to_watch = [
+            Path.cwd() / 'content',
+            Path.cwd() / 'layout',
+        ]
+
+        patterns_to_watch = ["*.py", "*.js", "*.css", "*.md", "*.html"]  # Only watch these file types
+
+        # Example: Patterns to ignore
+        patterns_to_ignore = [
+            "*.pyc",  # Ignore Python cache files
+            "*.log",  # Ignore log files
+            "*/node_modules/*",  # Ignore node_modules directory
+            "*/.git/*",  # Ignore git directory
+            "*/venv/*",  # Ignore virtual environment
+            "*/__pycache__/*"  # Ignore Python cache directories
+        ]
+
+        callback_with_partial = lambda event: self.build(event.src_path)
+
+        watch_multiple_directories(
+            directories_to_watch,
+            callback_with_partial,
+            patterns=patterns_to_watch,
+            ignore_patterns=patterns_to_ignore
+        )
